project/requirements.txt
allure-pytest
python-dotenv
dataclasses_json
pytest==7.0.1
selenium==4.1.3
webdriver-manager==3.5.3
Faker==13.2.0
requests==2.27.1
urllib3==1.26.8
pytest-xdist==2.5.0
jinja2<3.1.0
itsdangerous==2.0.1
flask==2.1.0
werkzeug==2.1.1
pymysql==1.0.2
sqlalchemy==1.4.31
pytest-xdist==2.5.0

